% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{flightGraph}
\alias{flightGraph}
\title{Imaginary flight network}
\format{An igraph graph with 20 nodes and 192 edges
\itemize{
  \item node: An origin or destination city
  \item edge: A flight between two cities
}}
\usage{
flightGraph
}
\description{
Network graph of an imaginary airline's flight routes.
}
\keyword{datasets}

