% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{flightLayout}
\alias{flightLayout}
\title{Plotting layout for nodes of a network about flight data}
\format{A matrix of two columns and 20 rows
\itemize{
  \item column 1: X position
  \item column 2: X position
}}
\usage{
flightLayout
}
\description{
Plotting layout for nodes of a network about flight data
}
\keyword{datasets}

