% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraphConversion.r
\name{igraphToGexf}
\alias{igraphToGexf}
\title{igraphToGexf}
\usage{
igraphToGexf(g)
}
\arguments{
\item{g}{An igraph object}
}
\value{
A string of json text representing the igraph object as an array holding an array each for nodes and edges.
}
\description{
Convert an igraph object to json
}
\details{
Converts an igraph object to json with an array holding an array each for nodes and edges.
}
\examples{
data(flightGraph)
igraphToGexf(flightGraph)

}
\author{
Jared P. Lander
}

