% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraphConversion.r
\name{prepIgraphConversion}
\alias{prepIgraphConversion}
\title{prepIgraphConversion}
\usage{
prepIgraphConversion(g)
}
\arguments{
\item{g}{An igraph object}
}
\value{
A list holding a nodes data.frames and an edges data.frame.  The nodes data.frame has at minimum an id and label column plus any other node attributes.  The edges data.frame has at minimum an id, source and target columns plus any edge attributes.
}
\description{
Prepares a graph for storing in standard format
}
\details{
Breaks an igraph into a list holding a nodes data.frames and an edges data.frame.  The nodes data.frame has at minimum an id and label column plus any other node attributes.  The edges data.frame has at minimum an id, source and target columns plus any edge attributes.
}
\examples{

data(flightGraph)
prepIgraphConversion(flightGraph)
}
\author{
Jared P. Lander
}

